pyproject.toml
[project]
name = "sf-bop-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bs4>=0.0.2",
    "dbt-core>=1.9.1",
    "dbt-snowflake>=1.9.0",
    "dlt[duckdb,mssql,snowflake]>=1.5.0",
    "duckdb>=1.1.3",
    "fabric>=3.2.2",
    "ibm-db-sa>=0.4.1",
    "pandas>=2.2.3",
    "polars>=1.19.0",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.0.1",
    "pyvis>=0.3.2",
    "snowflake-snowpark-python>=1.26.0",
    "streamlit>=1.41.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.0",
    "sqlfluff>=3.3.0",
    "sqlfmt>=0.0.2",
    "wheel>=0.45.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.workspace]
members = ["bop_app"]

[tool.hatch.build.targets.wheel]
packages = ["sf_bop_data", "down"]
